==> Code/1_Cleaning.Rmd <==
---
title: "1_Cleaning"
output: pdf_document
---

```{r}

library(ggplot2)
library(dplyr)
library(tidyr)
library(readr)

```

```{r}

# LOADING IN DATA

# Load in election outcome data
load("~/Thesis_Cook/Data/Raw/Presidential_Outcomes_1868_2020.Rdata")
elec_results <- pres_elections_release

# Read in fixed-width text file for demographic data
fwf_spec <- fwf_widths(
  widths = c(4, 2, 2, 3, 2, 1, 1, 1, 2, 8),
  col_names = c("Year", "State_Postal_Abbreviation", "State_FIPS_Code", "County_FIPS_Code", "Registry", "Race", "Origin", "Sex", "Age", "Population")
)

file_path <- "~/Thesis_Cook/Data/Raw/SEER_1969_2022.txt"
demographics <- read_fwf(file_path, fwf_spec, col_types = "iciiiiiiii")

```

```{r}

# TREATING MISSING VALUES

# Subset data (since demographic data is 1969 - 2020)
elec_results_subset <- subset(elec_results, election_year >= 1969)

# Check for missing vales
missing_values_dem <- colSums(is.na(demographics))
missing_values_elec <- colSums(is.na(elec_results_subset))

print(missing_values_dem) # None!
print(missing_values_elec)

# Handle missing rows for: democratic_raw_votes, republican_raw_votes, pres_raw_vote_totals_two_party, raw_county_vote_totals
missing_rows <- elec_results_subset[is.na(elec_results_subset$democratic_raw_votes), ]
print(missing_rows) 
print(elec_results_subset[elec_results_subset$county_name == "KALAWAO", ])

# KAWALAO county contains all of the NA rows for: democratic_raw_votes, republican_raw_votes, pres_raw_vote_totals_two_party, raw_county_vote_totals, so we will remove the county (also has very small population)
elec_results_subset <- elec_results_subset[elec_results_subset$county_name != "KALAWAO", ]

missing_values_elec <- colSums(is.na(elec_results_subset)) # Check to ensure there are no more NAs
print(missing_values_elec) # None! (that are relevant)

```

```{r}

# FEATURE ENGINEERING - DEMOGRAPHICS

print(unique(demographics$Race))
print(unique(demographics$Sex))
print(unique(demographics$Age))

# Instead of 18 different age groups, I will have the following breakdown:
# (1: 15-24), (2: 25-39), (3: 40-54), (4:55-69), (5: 70+)

demographics <- demographics %>%
  mutate(Age_Group = case_when(
    Age %in% c(0, 1, 2, 3) ~ 0,
    Age %in% c(4, 5, 6) ~ 1,
    Age %in% c(6, 7, 8) ~ 2,
    Age %in% c(9, 10, 11) ~ 3,
    Age %in% c(12, 13, 14) ~ 4,
    Age %in% c(15, 16, 17, 18) ~ 5
  ))

demographics <- demographics %>%
  mutate(
    sex_dummy = ifelse(Sex == 1, "Male", "Female"),
    race_dummy = case_when(
      Race == 1 ~ "White",
      Race == 2 ~ "Black",
      Race == 3 ~ "Other"
    ),
    Demo_Group = paste0(sex_dummy, "_", race_dummy, "_", Age_Group)
  ) %>%
  select(-sex_dummy, -race_dummy) 

# Dropping origin column (not relevant to this data)
demographics <- demographics %>% 
  select(-Origin)

# Create 5 digit fips code for demographics
demographics$fips <- sprintf("%02d%03d", demographics$State_FIPS_Code, demographics$County_FIPS_Code)

```

```{r}

# RESHAPING - DEMOGRAPHICS

# Sum populations across demographic groups
demographics_group <- demographics %>%
  group_by(Year, fips, Demo_Group) %>%
  summarise(Population = sum(Population, na.rm = TRUE), .groups = "drop")
  
# Widen data so demographic groups become columns
demographics_wide <- demographics_group %>% 
  pivot_wider(
    names_from = Demo_Group,
    values_from = Population
  )

# Change NAs to 0
demographics_wide[is.na(demographics_wide)] <- 0

# Rearrange
demographics_wide <- demographics_wide %>% 
  select(1,2,3,4,5,6,7,8,9,33,10,11,12,34,13,14,15,16,17,18,19,20,21,22,23,24,25,35,36,37,26,38,27,28,29,30,31,32)

# Check results
print(sum(demographics_group$Population) == sum(demographics$Population))

demographics_test <- demographics_wide %>% 
  select(-Year, -fips) %>% 
  mutate(total_pop = rowSums(across(where(is.numeric))))

print(sum(demographics$Population) == sum(demographics_test$total_pop))

# Adding in population totals for age, sex, and race
demographics_wide <- demographics_wide %>%
  mutate(Black_Total_Pop = rowSums(select(., matches("^(Female|Male)_Black_"))),
         White_Total_Pop = rowSums(select(., matches("^(Female|Male)_White_"))),
         Other_Total_Pop = rowSums(select(., matches("^(Female|Male)_Other_"))),
         Female_Total_Pop = rowSums(select(., starts_with("Female_"))),
         Male_Total_Pop = rowSums(select(., starts_with("Male_"))),
         Female_White_Total_Pop = rowSums(select(., matches("^(Female)_White_"))),
         Female_Black_Total_Pop = rowSums(select(., matches("^(Female)_Black_"))),
         Female_Other_Total_Pop = rowSums(select(., matches("^(Female)_Other_"))),
         Male_White_Total_Pop = rowSums(select(., matches("^(Male)_White_"))),
         Male_Black_Total_Pop = rowSums(select(., matches("^(Male)_Black_"))),
         Male_Other_Total_Pop = rowSums(select(., matches("^(Male)_Other_"))),
         Age_0_Total_Pop = rowSums(select(., ends_with("_0"))),
         Age_1_Total_Pop = rowSums(select(., ends_with("_1"))),
         Age_2_Total_Pop = rowSums(select(., ends_with("_2"))),
         Age_3_Total_Pop = rowSums(select(., ends_with("_3"))),
         Age_4_Total_Pop = rowSums(select(., ends_with("_4"))),
         Age_5_Total_Pop = rowSums(select(., ends_with("_5")))) %>% 
  mutate(Total_Pop = Female_Total_Pop + Male_Total_Pop)

# Check results
pop_by_race <- demographics_wide %>% 
  select(Black_Total_Pop, White_Total_Pop, Other_Total_Pop) %>% 
  sum()

pop_by_sex <- demographics_wide %>% 
  select(Female_Total_Pop, Male_Total_Pop) %>% 
  sum()

pop_by_age <- demographics_wide %>% 
  select(Age_0_Total_Pop, Age_1_Total_Pop, Age_2_Total_Pop, Age_3_Total_Pop, Age_4_Total_Pop, Age_5_Total_Pop) %>% 
  sum()

print(all(pop_by_race, pop_by_sex, pop_by_age, sum(demographics$Population), sum(demographics_wide$Total_Pop)))

```

```{r}

# MERGING

elec_results_subset <- elec_results_subset %>% 
  select(-county_first_date, -county_end_date, -state_admission_date, -original_county_name, -original_name_end_date)

# Check for missing values post-merge
merged_data <- elec_results_subset %>%
  left_join(demographics_wide, by = c("election_year" = "Year", "fips" = "fips"))

missing_values_merged <- colSums(is.na(merged_data))
print(missing_values_merged) 

missing_rows_merged <- merged_data[is.na(merged_data$Total_Pop), ]
print(missing_rows_merged) # All NAs due to demographic data not going all the way back to 1972 (although election data does)

# Drop NAs (so panel data is balanced)
merged_data <- merged_data %>%
  filter(complete.cases(Total_Pop))

# Check to make sure merged data is now balanced
missing_values_merged <- colSums(is.na(merged_data))
print(missing_values_merged) # None!

```

```{r}

# ADDITIONAL FEATURE ENGINEERING

cols_to_convert <- c("Female_Black_0", "Female_Black_1", "Female_Black_2", "Female_Black_3", "Female_Black_4", "Female_Black_5", "Female_Other_0", "Female_Other_1", "Female_Other_2", "Female_Other_3", "Female_Other_4", "Female_Other_4", "Female_Other_5", "Female_White_0", "Female_White_1", "Female_White_2", "Female_White_3", "Female_White_4", "Female_White_5", "Male_Black_0", "Male_Black_1", "Male_Black_2", "Male_Black_3", "Male_Black_4", "Male_Black_5", "Male_Other_0", "Male_Other_1", "Male_Other_2", "Male_Other_3", "Male_Other_4" ,"Male_Other_5", "Male_White_0", "Male_White_1", "Male_White_2", "Male_White_3", "Male_White_4", "Male_White_5", "Black_Total_Pop", "White_Total_Pop", "Other_Total_Pop", "Female_Total_Pop", "Male_Total_Pop", "Female_White_Total_Pop", "Female_Black_Total_Pop", "Female_Other_Total_Pop", "Male_White_Total_Pop", "Male_Black_Total_Pop", "Male_Other_Total_Pop", "Age_0_Total_Pop", "Age_1_Total_Pop", "Age_2_Total_Pop", "Age_3_Total_Pop", "Age_4_Total_Pop", "Age_5_Total_Pop")

merged_data <- merged_data %>%
  mutate(across(
    all_of(cols_to_convert),
    ~ .x / Total_Pop
  ))

merged_data <- merged_data %>% 
  mutate(dem_vote_share_two_party = democratic_raw_votes / pres_raw_county_vote_totals_two_party,
         dem_vote_share_total = democratic_raw_votes / raw_county_vote_totals,
         rep_vote_share_two_party = republican_raw_votes / pres_raw_county_vote_totals_two_party,
         rep_vote_share_total = republican_raw_votes / raw_county_vote_totals,
         dem_voter_turnout = democratic_raw_votes / Total_Pop,
         rep_voter_turnout = republican_raw_votes / Total_Pop,
         total_voter_turnout = raw_county_vote_totals / Total_Pop)

merged_data <- merged_data %>% 
  select(1,2,3,4,5,6,7,8,9,10,11,12,13,14,70,71,72,73,74,75,76,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69)

# Save data as rds file
saveRDS(merged_data, file = "../Data/Processed/Merged_Data.rds")

```
