==> Code/3B_Models.Rmd <==
---
title: "3B_Models"
output: pdf_document
---

```{r}

library(ggplot2)
library(dplyr)
library(tidyr)
library(readr)
library(sf)
library(tigris)
library(plm)
library(lmtest)
library(lme4)
library(glmnet)
library(kableExtra)
library(modelr)
library(car)
library(lmtest)
library(sandwich)
library(gt)
library(broom)

```

```{r}

# BUILDING TRAIN & TEST (70/30)

data <- readRDS("../Data/Processed/Merged_Data.rds")
model_data <- data %>% 
  select(election_year, fips, dem_vote_share_two_party, Female_Total_Pop, Male_Total_Pop, White_Total_Pop, Black_Total_Pop, Other_Total_Pop, Age_0_Total_Pop, Age_1_Total_Pop, Age_2_Total_Pop, Age_3_Total_Pop, Age_4_Total_Pop, Age_5_Total_Pop, Total_Pop)

train <- model_data[!(model_data$election_year %in% c(1972, 1976, 2016, 2020)), ]
test_early <- model_data[model_data$election_year %in% c(1972, 1976), ]
test_late <- model_data[model_data$election_year %in% c(2016, 2020), ]

pdata <- pdata.frame(model_data, index = c("fips", "election_year"))
pdata_train <- pdata.frame(train, index = c("fips", "election_year"))
pdata_test_early <- pdata.frame(test_early, index = c("fips", "election_year"))
pdata_test_late <- pdata.frame(test_late, index = c("fips", "election_year"))

```

```{r}

# MODEL SEQUENCE: BASE MODELS

# M1A: Age Only (OLS)
m_age <- plm(dem_vote_share_two_party ~ Age_0_Total_Pop + Age_1_Total_Pop + Age_2_Total_Pop + Age_4_Total_Pop + Age_5_Total_Pop, data = pdata_train, model = "pooling")

# M1B: Sex Only (OLS)
m_sex <- plm(dem_vote_share_two_party ~ Female_Total_Pop, data = pdata_train, model = "pooling")

# M1C: Race Only (OLS)
m_race <- plm(dem_vote_share_two_party ~ Black_Total_Pop + White_Total_Pop, data = pdata_train, model = "pooling")

model <- dem_vote_share_two_party ~ Age_0_Total_Pop + Age_1_Total_Pop + Age_2_Total_Pop + Age_4_Total_Pop + Age_5_Total_Pop + Female_Total_Pop + Black_Total_Pop + White_Total_Pop

# M2: All Main Effects (Age + Sex + Race)
m_pooled <- plm(model, data = pdata_train, model = "pooling")

# Linear Model (same as plm, but useful for different function calls)
lm_fit <- lm(model, data = pdata_train)
#crPlots(lm_fit)

```

```{r}

# BASE MODELS: COMPARISON

# Summary Statistics
summary(m_age)
summary(m_sex)
summary(m_race)
summary(m_pooled)

# F-Tests: Model Complexity
pFtest(m_pooled, m_age)
pFtest(m_pooled, m_sex)
pFtest(m_pooled, m_race) # Result for all: Complex models perform better

# R^2 Analysis
r2_values_base <- c(
  m_age = summary(m_age)$r.squared,
  m_sex = summary(m_sex)$r.squared,
  m_race = summary(m_race)$r.squared,
  m_pooled = summary(m_pooled)$r.squared
)

r2_df_base <- data.frame(
  model = names(r2_values_base),
  r_squared = as.numeric(r2_values_base)
)

r2_adj_df_base <- 
  r2_df_base[r2_df_base$model %in% c("m_age.adjrsq", "m_sex.adjrsq", "m_race.adjrsq", "m_pooled.adjrsq"), ] %>% 
  mutate(across(where(is.numeric), ~ round(.x, 2)))

model_names_base <- c("Age Model", "Sex Model", "Race Model", "Age-Sex-Race Model")
r2_adj_df_base$model <- model_names_base
r2_adj_df_base$model <- factor(r2_adj_df_base$model, levels = r2_adj_df_base$model)

r2_base_models <- ggplot(r2_adj_df_base, aes(x = model, y = r_squared)) +
  geom_point() +
  labs(x = "Model", y = "Adjusted R Squared", title = "Adjusted R Squared by Model Type - Base Models") +
  theme_minimal()

print(r2_base_models)
print(r2_adj_df_base %>% kable())
print(summary(m_pooled)) # Model of interest (age-sex-race model)

```

```{r}

# MODEL SEQUENCE: COUNTY & YEAR EFFECTS

# M3: County Effects
m_fips <- lm(dem_vote_share_two_party ~ factor(fips), data = pdata_train)

# M4: Year Effects
m_year <- lm(dem_vote_share_two_party ~ factor(election_year), data = pdata_train)

# M5: County & Year Effects
m_year_fips <- lm(dem_vote_share_two_party ~ factor(fips) + factor(election_year), data = pdata_train)

r2_values_year_fips <- c(
  m_fips = summary(m_fips)$r.squared,
  m_year = summary(m_year)$r.squared,
  m_year_fips = summary(m_year_fips)$r.squared)

r2_df_year_fips <- data.frame(
  model = names(r2_values_year_fips),
  r_squared = as.numeric(r2_values_year_fips)) %>% 
  mutate(across(where(is.numeric), ~ round(.x, 2)))

model_names_year_fips <- c("County Effects Model", "Year Effects Model Model", "County-Year Effects Model")
r2_df_year_fips$model <- model_names_year_fips
r2_df_year_fips$model <- factor(r2_df_year_fips$model, levels = r2_df_year_fips$model)

r2_year_fips_models <- ggplot(r2_df_year_fips, aes(x = model, y = r_squared)) +
  geom_point() +
  labs(x = "Model", y = "Adjusted R Squared", title = "Adjusted R Squared by Model Type - County & Year Effects") +
  theme_minimal()

print(r2_year_fips_models)
print(r2_df_year_fips %>% kable())

```

```{r}

# MODEL SEQUENCE: FIXED & RANDOM EFFECTS

# M6A: Year FE
m_year_fe <- plm(model, data = pdata_train, model = "within", effect = "time")
 
# M6B: County FE
m_county_fe <- plm(model, data = pdata_train, model = "within", effect = "individual")

# M7: Two-Way FE
m_tw_fe <- plm(model, data = pdata_train, model = "within", effect = "twoways")

# M8: RE
m_re <- plm(model, data = pdata_train, model = "random")

```

```{r}

# MODEL TESTING

# Pooled OLS can have much higher R^2 because it falsely attributes fixed traits such as county partisanship to the included variables, whereas TWFE drops that explanatory power (OLS can be more biased)

# TWFE: The partial effect of a one-unit change in the demographic variable on Democratic two-party vote share, holding constant unobserved county characteristics and year-specific shocks

# F-Test (Fixed Effects vs. OLS)
pFtest(m_year_fe, m_pooled) # Prefer FE
pFtest(m_county_fe, m_pooled) # Prefer FE

# Breusch-Pagan Lagrange Multiplier Test (Random Effects vs. OLS)
plmtest(m_pooled, type = "bp") # Prefer RE

# Hausman Test (Fixed vs. Random Effects)
phtest(m_year_fe, m_re) # Prefer year FE
phtest(m_county_fe, m_re) # Prefer county FE
phtest(m_tw_fe, m_re) # Prefer TWFE

```

```{r}

# Clustered SEs by county
clustered_se <- vcovHC(m_tw_fe, type = "HC1", cluster = "group")

# Tidy with clustered SEs
model_tidy <- broom::tidy(coeftest(m_tw_fe, vcov = clustered_se))

# 2. Create the Coefficient Plot
ggplot(model_tidy, aes(x = estimate, y = term)) +
  geom_point() +
  geom_errorbarh(aes(xmin = estimate - 1.96*std.error, xmax = estimate + 1.96*std.error), height = 0.2) +
  geom_vline(xintercept = 0, linetype = "dashed") +
  labs(
    title = "Estimated Effects of Demographic Composition on Democratic Vote Share",
    x = "Coefficient Estimate",
    y = "Demographic Variables"
  ) +
  theme_minimal()

```

```{r}

# MODEL COMPARISON

#plm(dem_vote_share_two_party ~ election_year, pdata_train)

# R^2 Analysis
r2_values <- c(
  m_year_fe = summary(m_year_fe)$r.squared,
  m_county_fe = summary(m_county_fe)$r.squared,
  m_tw_fe = summary(m_tw_fe)$r.squared,
  m_re = summary(m_re)$r.squared,
  m_pooled = summary(m_pooled)$r.squared
)

r2_df <- data.frame(
  model = names(r2_values),
  r_squared = as.numeric(r2_values)
)

r2_adj_df <- 
  r2_df[r2_df$model %in% c("m_year_fe.adjrsq", "m_county_fe.adjrsq", "m_tw_fe.adjrsq", "m_re.adjrsq", "m_pooled.adjrsq"),] %>% 
  mutate(across(where(is.numeric), ~ round(.x, 2)))

model_names <- c("Year FE", "County FE", "TWFE", "Random Effects", "Pooled OLS")
r2_adj_df$model <- model_names
r2_adj_df$model <- factor(r2_adj_df$model, levels = r2_adj_df$model)

r2_models <- ggplot(r2_adj_df, aes(x = model, y = r_squared)) +
  geom_point() +
  labs(x = "Model", y = "Adjusted R Squared", title = "Adjusted R^2 by Model Type") +
  theme_minimal()

print(r2_models)
print(r2_adj_df %>% kable())

```

```{r}

# PREDICTION ON TEST SETS

X_early <- model.matrix(model, data = pdata_test_early)[,-1]
X_late <- model.matrix(model, data = pdata_test_late)[,-1]

coefs_pooled <- coef(m_pooled)
coefs_year_fe <- coef(m_year_fe)
coefs_county_fe <- coef(m_county_fe)
coefs_re <- coef(m_re)
coefs_fips <- coef(m_fips)
coefs_year <- coef(m_year)
coefs_year_fips <- coef(m_year_fips)

# For year/county FE, predictions are deviations (not actual vote share levels)
# Given the average vote share for this year, how does this county's demographics explain its relative deviation?

pred_early_year_fe <- X_early %*% coefs_year_fe 
pred_late_year_fe  <- X_late  %*% coefs_year_fe

pred_early_county_fe <- X_early %*% coefs_county_fe
pred_late_county_fe  <- X_late  %*% coefs_county_fe

pred_early_re <- X_early %*% coefs_re[-1]
pred_late_re  <- X_late  %*% coefs_re[-1]

pred_early_pooled <- X_early %*% coefs_pooled[-1]
pred_late_pooled  <- X_late  %*% coefs_pooled[-1]

pred_early_fips <- predict(m_fips, newdata = pdata_test_early)
pred_late_fips <- predict(m_fips, newdata = pdata_test_late)

```

```{r}

# COMPARISON OF PREDICTIONS

rmse <- function(pred, actual) {
  sqrt(mean((pred - actual)^2))
}

mae <- function(pred, actual) {
  mean(abs(actual - pred))
}

# Year FE
mae_early_year_fe <- mae(pred_early_year_fe, pdata_test_early$dem_vote_share_two_party)
mae_late_year_fe <- mae(pred_late_year_fe, pdata_test_late$dem_vote_share_two_party)

# County FE
mae_early_county_fe <- mae(pred_early_county_fe, pdata_test_early$dem_vote_share_two_party)
mae_late_county_fe <- mae(pred_late_county_fe, pdata_test_late$dem_vote_share_two_party)

rmse_early_county_fe <- rmse(pred_early_county_fe, pdata_test_early$dem_vote_share_two_party)
rmse_late_county_fe <- rmse(pred_late_county_fe, pdata_test_late$dem_vote_share_two_party)

mae_change_county_fe <- (mae_late_county_fe - mae_early_county_fe) / (mae_early_county_fe) * 100
rmse_change_county_fe <- (rmse_late_county_fe - rmse_early_county_fe) / (rmse_early_county_fe) * 100

# RE
mae_early_re <- mae(pred_early_re, pdata_test_early$dem_vote_share_two_party)
mae_late_re <- mae(pred_late_re, pdata_test_late$dem_vote_share_two_party)

rmse_early_re <- rmse(pred_early_re, pdata_test_early$dem_vote_share_two_party)
rmse_late_re <- rmse(pred_late_re, pdata_test_late$dem_vote_share_two_party)

mae_change_re <- (mae_late_re - mae_early_re) / (mae_early_re) * 100
rmse_change_re <- (rmse_late_re - rmse_early_re) / (rmse_early_re) * 100

# Pooled OLS
mae_early_pooled <- mae(pred_early_pooled, pdata_test_early$dem_vote_share_two_party)
mae_late_pooled <- mae(pred_late_pooled, pdata_test_late$dem_vote_share_two_party)

rmse_early_pooled <- rmse(pred_early_pooled, pdata_test_early$dem_vote_share_two_party)
rmse_late_pooled <- rmse(pred_late_pooled, pdata_test_late$dem_vote_share_two_party)

mae_change_pooled <- (mae_late_pooled - mae_early_pooled) / (mae_early_pooled) * 100
rmse_change_pooled <- (rmse_late_pooled - rmse_early_pooled) / (rmse_early_pooled) * 100

# County Effect
mae_early_fips <- mae(pred_early_fips, pdata_test_early$dem_vote_share_two_party)
mae_late_fips <- mae(pred_late_fips, pdata_test_late$dem_vote_share_two_party)

rmse_early_fips <- rmse(pred_early_fips, pdata_test_early$dem_vote_share_two_party)
rmse_late_fips <- rmse(pred_late_fips, pdata_test_late$dem_vote_share_two_party)

mae_change_fips <- (mae_late_fips - mae_early_fips) / (mae_early_fips) * 100
rmse_change_fips <- (rmse_late_fips - rmse_early_fips) / (rmse_early_fips) * 100

# Comparison
eval_comparison <- data.frame(
  Model = c("Pooled OLS", "Random Effects", "County Fixed Effects", "County Effects"),
  MAE_Early = c(mae_early_pooled, mae_early_re, mae_early_county_fe, mae_early_fips),
  MAE_Late  = c(mae_late_pooled,  mae_late_re, mae_late_county_fe, mae_late_fips),
  MAE_Perc_Change = c(mae_change_pooled, mae_change_re, mae_change_county_fe, mae_change_fips),
  RMSE_Early = c(rmse_early_pooled, rmse_early_re, rmse_early_county_fe, rmse_early_fips),
  RMSE_Late  = c(rmse_late_pooled, rmse_late_re, rmse_late_county_fe, rmse_late_fips),
  RMSE_Perc_Change = c(rmse_change_pooled, rmse_change_re, rmse_change_county_fe, rmse_change_fips)
)

print(eval_comparison)

#plm(dem_vote_share_two_party ~ as.numeric(election_year), pdata_train)


```

```{r}

# RANDOM FOREST

library(randomForest)

m_rf <- randomForest(model, data = pdata_train, ntree = 500)

pred_rf_early <- predict(m_rf, newdata = pdata_test_early)
pred_rf_late  <- predict(m_rf, newdata = pdata_test_late)

mae_early_rf<- mae(pred_rf_early, pdata_test_early$dem_vote_share_two_party)
mae_late_rf <- mae(pred_rf_late, pdata_test_late$dem_vote_share_two_party)

rmse_early_rf <- rmse(pred_rf_early, pdata_test_early$dem_vote_share_two_party)
rmse_late_rf <- rmse(pred_rf_late, pdata_test_late$dem_vote_share_two_party)

mae_change_rf <- (mae_late_rf - mae_early_rf) / (mae_early_rf) * 100
rmse_change_rf <- (rmse_late_rf - rmse_early_rf) / (rmse_early_rf) * 100

eval_comparison_rf <- data.frame(
  Model = c("Pooled OLS", "Random Effects", "County Fixed Effects", "County Effects", "Random Forest"),
  MAE_Early = c(mae_early_pooled, mae_early_re, mae_early_county_fe, mae_early_fips, mae_early_rf),
  MAE_Late  = c(mae_late_pooled,  mae_late_re, mae_late_county_fe, mae_late_fips, mae_late_rf),
  MAE_Perc_Change = c(mae_change_pooled, mae_change_re, mae_change_county_fe, mae_change_fips, mae_change_rf),
  RMSE_Early = c(rmse_early_pooled, rmse_early_re, rmse_early_county_fe, rmse_early_fips, rmse_early_rf),
  RMSE_Late  = c(rmse_late_pooled, rmse_late_re, rmse_late_county_fe, rmse_late_fips, rmse_late_rf),
  RMSE_Perc_Change = c(rmse_change_pooled, rmse_change_re, rmse_change_county_fe, rmse_change_fips, rmse_change_rf)
)

print(eval_comparison_rf)

```

```{r}

# SUMMARY CHARTS & TABLES

pdata_test_early <- pdata_test_early %>% 
  mutate(pooled_prediction = pred_early_pooled,
         re_prediction = pred_early_re,
         rf_prediction = pred_rf_early,
         county_fe_prediction = pred_early_county_fe,
         fips_prediction = pred_early_fips,
         test_period = "Early")

pdata_test_late <- pdata_test_late %>% 
  mutate(pooled_prediction = pred_late_pooled,
         re_prediction = pred_late_re,
         rf_prediction = pred_rf_late,
         county_fe_prediction = pred_late_county_fe,
         fips_prediction = pred_late_fips,
         test_period = "Late")

pdata_test <- rbind(pdata_test_early, pdata_test_late)

pooled_comparison <- ggplot(pdata_test, aes(x = pooled_prediction, y = dem_vote_share_two_party, color = test_period)) +
  geom_point() +
  geom_abline(slope = 1)

re_comparison <- ggplot(pdata_test, aes(x = re_prediction, y = dem_vote_share_two_party, color = test_period)) +
  geom_point() +
  geom_abline(slope = 1)

rf_comparison <- ggplot(pdata_test, aes(x = rf_prediction, y = dem_vote_share_two_party, color = test_period)) +
  geom_point() +
  geom_abline(slope = 1)

county_fe_comparison <- ggplot(pdata_test, aes(x = county_fe_prediction, y = dem_vote_share_two_party, color = test_period)) +
  geom_point() +
  geom_abline(slope = 1)

fips_comparison <- ggplot(pdata_test, aes(x = fips_prediction, y = dem_vote_share_two_party, color = test_period)) +
  geom_point() + 
  geom_abline(slope = 1)

print(pooled_comparison)
print(re_comparison)
print(rf_comparison)
print(county_fe_comparison)
print(fips_comparison)

```

```{r}

# SAVING MODELS

save(
  m_age,
  m_sex,
  m_race,
  m_pooled,
  m_fips,
  m_year,
  m_year_fips,
  m_county_fe,
  m_year_fe,
  m_tw_fe,
  m_re,
  m_rf,
  file = "Models.RData"
)

```
